Move log linking to the log itself
