+++ Assets/Scripts/Incidents/IncidentReport.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Text.RegularExpressions;
 
 namespace Game.Incidents
 {
@@ -39,6 +40,24 @@ namespace Game.Incidents
 			}
 
 			ReportLog = fullLog;
+		}
+
+		public string GenerateLinkedLog()
+		{
+			var textLine = string.Copy(ReportLog);
+			textLine = FlavorService.Instance.GenerateFlavor(textLine);
+
+			var matches = Regex.Matches(textLine, @"\{(\d+)\}");
+
+			foreach (Match match in matches)
+			{
+				var matchString = match.Value;
+				var linkedContext = Contexts[matchString];
+				var linkString = string.Format("<link=\"{0}\">{1}</link>", linkedContext.ID, linkedContext.Name);
+				textLine = textLine.Replace(matchString, linkString);
+			}
+
+			return textLine;
 		}
 	}
 }
